requirements.txt
# Core ML libraries
tensorflow==2.13.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
Pillow>=8.0.0
opencv-python>=4.5.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Jupyter and development
jupyter>=1.0.0
notebook>=6.0.0

# Utilities
tqdm>=4.0.0
requests>=2.25.0

# Ground station dependencies
folium>=0.12.0
pyserial>=3.0

# Optional: For advanced features
# torch>=1.9.0  # If using PyTorch for comparison
# torchvision>=0.10.0
